requirements.txt
# PyTorch 관련
torch==1.13.1
torchvision==0.14.1
torchaudio==0.13.1

# 기본 데이터 과학 라이브러리 (미리 컴파일된 버전 사용)
numpy>=1.21.0
pandas>=1.3.0
pillow>=10.0.0
opencv-python>=4.5.0

# 시각화
matplotlib>=3.4.0
seaborn>=0.11.0

# 머신러닝
scikit-learn>=1.0.0

# 진행 표시
tqdm>=4.62.0

# 기타
scipy>=1.7.0
boto3>=1.20.0

torchsummary==1.5.1

huggingface_hub>=0.22.0
safetensors>=0.4.0

torchxrayvision>=0.0.45
timm>=0.9.0  # (일부 환경에서 XRV가 timm에 의존)